+++ tobs_blaine/fxmanifest.lua
@@ -3,7 +3,7 @@ game "gta5"
 
 author "Vega"
 description "Paleto and Fleeca bank heists for ESX"
-version "1.4.1"
+version "1.5.0"
 
 shared_scripts {"config/config.lua", "locales/locales.lua"}
 client_scripts {"client/bridge.lua", "client/main.lua"}